requirements.txt
flask
flask-bootstrap
flask-wtf
flask-testing
flask-login
python-dotenv
openai
gunicorn
pymongo
requests